server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Werkzeug==3.1.3
python-dotenv==1.0.1
pdfplumber==0.11.5
python-docx==1.1.2
anthropic==0.51.0
razorpay==1.4.2
gunicorn==23.0.0
itsdangerous==2.2.0
reportlab==4.2.5
psycopg[binary]
Flask-Limiter==4.1.0
sendgrid==6.11.0

setuptools
